move _PackageBoundObject into Scaffold
