+++ src/flask/helpers.py
@@ -1,13 +1,10 @@
 import os
-import pkgutil
 import socket
-import sys
 import warnings
 from functools import update_wrapper
 from threading import RLock
 
 import werkzeug.utils
-from jinja2 import FileSystemLoader
 from werkzeug.exceptions import NotFound
 from werkzeug.routing import BuildError
 from werkzeug.urls import url_quote
@@ -677,157 +674,6 @@ def send_from_directory(directory, path, **kwargs):
     )
 
 
-def get_root_path(import_name):
-    """Returns the path to a package or cwd if that cannot be found.  This
-    returns the path of a package or the folder that contains a module.
-
-    Not to be confused with the package path returned by :func:`find_package`.
-    """
-    # Module already imported and has a file attribute.  Use that first.
-    mod = sys.modules.get(import_name)
-    if mod is not None and hasattr(mod, "__file__"):
-        return os.path.dirname(os.path.abspath(mod.__file__))
-
-    # Next attempt: check the loader.
-    loader = pkgutil.get_loader(import_name)
-
-    # Loader does not exist or we're referring to an unloaded main module
-    # or a main module without path (interactive sessions), go with the
-    # current working directory.
-    if loader is None or import_name == "__main__":
-        return os.getcwd()
-
-    if hasattr(loader, "get_filename"):
-        filepath = loader.get_filename(import_name)
-    else:
-        # Fall back to imports.
-        __import__(import_name)
-        mod = sys.modules[import_name]
-        filepath = getattr(mod, "__file__", None)
-
-        # If we don't have a filepath it might be because we are a
-        # namespace package.  In this case we pick the root path from the
-        # first module that is contained in our package.
-        if filepath is None:
-            raise RuntimeError(
-                "No root path can be found for the provided module"
-                f" {import_name!r}. This can happen because the module"
-                " came from an import hook that does not provide file"
-                " name information or because it's a namespace package."
-                " In this case the root path needs to be explicitly"
-                " provided."
-            )
-
-    # filepath is import_name.py for a module, or __init__.py for a package.
-    return os.path.dirname(os.path.abspath(filepath))
-
-
-def _matching_loader_thinks_module_is_package(loader, mod_name):
-    """Given the loader that loaded a module and the module this function
-    attempts to figure out if the given module is actually a package.
-    """
-    cls = type(loader)
-    # If the loader can tell us if something is a package, we can
-    # directly ask the loader.
-    if hasattr(loader, "is_package"):
-        return loader.is_package(mod_name)
-    # importlib's namespace loaders do not have this functionality but
-    # all the modules it loads are packages, so we can take advantage of
-    # this information.
-    elif cls.__module__ == "_frozen_importlib" and cls.__name__ == "NamespaceLoader":
-        return True
-    # Otherwise we need to fail with an error that explains what went
-    # wrong.
-    raise AttributeError(
-        f"{cls.__name__}.is_package() method is missing but is required"
-        " for PEP 302 import hooks."
-    )
-
-
-def _find_package_path(root_mod_name):
-    """Find the path where the module's root exists in"""
-    import importlib.util
-
-    try:
-        spec = importlib.util.find_spec(root_mod_name)
-        if spec is None:
-            raise ValueError("not found")
-    # ImportError: the machinery told us it does not exist
-    # ValueError:
-    #    - the module name was invalid
-    #    - the module name is __main__
-    #    - *we* raised `ValueError` due to `spec` being `None`
-    except (ImportError, ValueError):
-        pass  # handled below
-    else:
-        # namespace package
-        if spec.origin in {"namespace", None}:
-            return os.path.dirname(next(iter(spec.submodule_search_locations)))
-        # a package (with __init__.py)
-        elif spec.submodule_search_locations:
-            return os.path.dirname(os.path.dirname(spec.origin))
-        # just a normal module
-        else:
-            return os.path.dirname(spec.origin)
-
-    # we were unable to find the `package_path` using PEP 451 loaders
-    loader = pkgutil.get_loader(root_mod_name)
-    if loader is None or root_mod_name == "__main__":
-        # import name is not found, or interactive/main module
-        return os.getcwd()
-    else:
-        if hasattr(loader, "get_filename"):
-            filename = loader.get_filename(root_mod_name)
-        elif hasattr(loader, "archive"):
-            # zipimporter's loader.archive points to the .egg or .zip
-            # archive filename is dropped in call to dirname below.
-            filename = loader.archive
-        else:
-            # At least one loader is missing both get_filename and archive:
-            # Google App Engine's HardenedModulesHook
-            #
-            # Fall back to imports.
-            __import__(root_mod_name)
-            filename = sys.modules[root_mod_name].__file__
-        package_path = os.path.abspath(os.path.dirname(filename))
-
-        # In case the root module is a package we need to chop of the
-        # rightmost part. This needs to go through a helper function
-        # because of namespace packages.
-        if _matching_loader_thinks_module_is_package(loader, root_mod_name):
-            package_path = os.path.dirname(package_path)
-
-    return package_path
-
-
-def find_package(import_name):
-    """Finds a package and returns the prefix (or None if the package is
-    not installed) as well as the folder that contains the package or
-    module as a tuple.  The package path returned is the module that would
-    have to be added to the pythonpath in order to make it possible to
-    import the module.  The prefix is the path below which a UNIX like
-    folder structure exists (lib, share etc.).
-    """
-    root_mod_name, _, _ = import_name.partition(".")
-    package_path = _find_package_path(root_mod_name)
-    site_parent, site_folder = os.path.split(package_path)
-    py_prefix = os.path.abspath(sys.prefix)
-    if package_path.startswith(py_prefix):
-        return py_prefix, package_path
-    elif site_folder.lower() == "site-packages":
-        parent, folder = os.path.split(site_parent)
-        # Windows like installations
-        if folder.lower() == "lib":
-            base_dir = parent
-        # UNIX like installations
-        elif os.path.basename(parent).lower() == "lib":
-            base_dir = os.path.dirname(parent)
-        else:
-            base_dir = site_parent
-        return base_dir, package_path
-    return None, package_path
-
-
 class locked_cached_property:
     """A decorator that converts a function into a lazy property.  The
     function wrapped is called the first time to retrieve the result
@@ -852,155 +698,6 @@ class locked_cached_property:
                 value = self.func(obj)
                 obj.__dict__[self.__name__] = value
             return value
-
-
-class _PackageBoundObject:
-    #: The name of the package or module that this app belongs to. Do not
-    #: change this once it is set by the constructor.
-    import_name = None
-
-    #: Location of the template files to be added to the template lookup.
-    #: ``None`` if templates should not be added.
-    template_folder = None
-
-    #: Absolute path to the package on the filesystem. Used to look up
-    #: resources contained in the package.
-    root_path = None
-
-    def __init__(self, import_name, template_folder=None, root_path=None):
-        self.import_name = import_name
-        self.template_folder = template_folder
-
-        if root_path is None:
-            root_path = get_root_path(self.import_name)
-
-        self.root_path = root_path
-        self._static_folder = None
-        self._static_url_path = None
-
-        # circular import
-        from .cli import AppGroup
-
-        #: The Click command group for registration of CLI commands
-        #: on the application and associated blueprints. These commands
-        #: are accessible via the :command:`flask` command once the
-        #: application has been discovered and blueprints registered.
-        self.cli = AppGroup()
-
-    @property
-    def static_folder(self):
-        """The absolute path to the configured static folder."""
-        if self._static_folder is not None:
-            return os.path.join(self.root_path, self._static_folder)
-
-    @static_folder.setter
-    def static_folder(self, value):
-        if value is not None:
-            value = os.fspath(value).rstrip(r"\/")
-        self._static_folder = value
-
-    @property
-    def static_url_path(self):
-        """The URL prefix that the static route will be accessible from.
-
-        If it was not configured during init, it is derived from
-        :attr:`static_folder`.
-        """
-        if self._static_url_path is not None:
-            return self._static_url_path
-
-        if self.static_folder is not None:
-            basename = os.path.basename(self.static_folder)
-            return f"/{basename}".rstrip("/")
-
-    @static_url_path.setter
-    def static_url_path(self, value):
-        if value is not None:
-            value = value.rstrip("/")
-
-        self._static_url_path = value
-
-    @property
-    def has_static_folder(self):
-        """This is ``True`` if the package bound object's container has a
-        folder for static files.
-
-        .. versionadded:: 0.5
-        """
-        return self.static_folder is not None
-
-    @locked_cached_property
-    def jinja_loader(self):
-        """The Jinja loader for this package bound object.
-
-        .. versionadded:: 0.5
-        """
-        if self.template_folder is not None:
-            return FileSystemLoader(os.path.join(self.root_path, self.template_folder))
-
-    def get_send_file_max_age(self, filename):
-        """Used by :func:`send_file` to determine the ``max_age`` cache
-        value for a given file path if it wasn't passed.
-
-        By default, this returns :data:`SEND_FILE_MAX_AGE_DEFAULT` from
-        the configuration of :data:`~flask.current_app`. This defaults
-        to ``None``, which tells the browser to use conditional requests
-        instead of a timed cache, which is usually preferable.
-
-        .. versionchanged:: 2.0
-            The default configuration is ``None`` instead of 12 hours.
-
-        .. versionadded:: 0.9
-        """
-        value = current_app.send_file_max_age_default
-
-        if value is None:
-            return None
-
-        return total_seconds(value)
-
-    def send_static_file(self, filename):
-        """Function used internally to send static files from the static
-        folder to the browser.
-
-        .. versionadded:: 0.5
-        """
-        if not self.has_static_folder:
-            raise RuntimeError("No static folder for this object")
-
-        # send_file only knows to call get_send_file_max_age on the app,
-        # call it here so it works for blueprints too.
-        max_age = self.get_send_file_max_age(filename)
-        return send_from_directory(self.static_folder, filename, max_age=max_age)
-
-    def open_resource(self, resource, mode="rb"):
-        """Opens a resource from the application's resource folder.  To see
-        how this works, consider the following folder structure::
-
-            /myapplication.py
-            /schema.sql
-            /static
-                /style.css
-            /templates
-                /layout.html
-                /index.html
-
-        If you want to open the :file:`schema.sql` file you would do the
-        following::
-
-            with app.open_resource('schema.sql') as f:
-                contents = f.read()
-                do_something_with(contents)
-
-        :param resource: the name of the resource.  To access resources within
-                         subfolders use forward slashes as separator.
-        :param mode: Open file in this mode. Only reading is supported,
-            valid values are "r" (or "rt") and "rb".
-        """
-        if mode not in {"r", "rt", "rb"}:
-            raise ValueError("Resources can only be opened for reading")
-
-        return open(os.path.join(self.root_path, resource), mode)
 
 
 def total_seconds(td):
